Rename CachedReadOnlyProperty CachingReadOnlyProperty and add license
